De agregan funciones a los modelos para filtrar

+++ app/Http/Controllers/PaymentsController.php
@@ -5,17 +5,19 @@ namespace App\Http\Controllers;
 use App\Models\Loans;
 use App\Models\Payments;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PaymentsController extends Controller
 {
     public function index(){
-        $payments = Payments::all();
-
+        $payments = Auth::user()->getPaymentsByPortafolio();
+        // dd($payments);
         return view('payment.index', compact('payments'));
     }
 
     public function create(){
-        $loans = Loans::all();
+        $user = Auth::user();
+        $loans = $user->getLoansByPortafolio;
         return view('payment.create', compact('loans'));
     }
 
@@ -26,7 +28,8 @@ class PaymentsController extends Controller
         if ($count + $request->amount <= $valueLoan->total_pay) {
             $payment = new Payments();
             $payment->loan_id = $request->loan_id;
-            $payment->amount =  (int)$request->amount;
+            $payment->amount =  $request->amount;
+            $payment->user_id =  Auth::user()->id;
             $payment->payment_date = $request->payment_date;
             $payment->save();
             return redirect()->route('payments')->with('status', 'Successfully created payment.');
@@ -45,9 +48,11 @@ class PaymentsController extends Controller
     }
 
     public function delete($id){
-        $this->decrypt($id);
-        $payment = Payments::find($id);
-        $payment->delete();
-        return redirect()->route('payments')->with('status', 'Successfully deleted payment.');
+        if (Auth::user()->hasRole('Admin')) {
+            $id = $this->decrypt($id);
+            $payment = Payments::find($id);
+            $payment->delete();
+            return redirect()->route('payments')->with('status', 'Successfully deleted payment.');
+        }
     }
 }
